+++ doc/CHANGES.rst
@@ -38,6 +38,10 @@ CHANGELOG
   use this function. 
 * [IMP] Add option **--anyblok-db-url** in plugin nose options. the options can have 
   default value from environment variable **ANYBLOK_DATABASE_URL**
+* [IMP] add primary_key attibute on Many2One and One2One.
+  Only when the column is created by the relationship Many2One or One2One.
+  The column created get the attribute primary_key gave to relationship and 
+  apply it.
 
 0.16.2 (2018-02-12)
 -------------------